improve galaxy test flow
